requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord==2.0.0
discord.py==2.0.1
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
StackAPI==0.2.0
urllib3==1.26.12
yarl==1.8.1
